Extend test suite to meet 80% coverage requirement (#148)
This commit adds more end-to-end tests for Image Based Video Search.
The intention is to cover >80% of the app functionality.
The update adds:

* Class methods to interact with the application
* Reusable test methods that make assertions on functionality
* Test methods composed into different suites

While testing, I bug was found with GPU config in compose
Fix is also included in this change set.

There are 4 test suites:

* Suite 1 do Recorded Stream Search with Input by Frame Capture
* Suite 2 do Live Stream Search with Input by Frame Capture
* Suite 3 do Recorded Stream Search with Input by Image Upload
* Suite 4 do Live Stream Search with Input by Image Upload

All the suites assert that the right number of results are displayed,
sorting functionality is working, and database clearing is effective.